src/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
SQLAlchemy==1.4.45
pydantic[email]~=1.10.13
envparse==0.2.0
psycopg2-binary==2.9.5
asyncpg==0.27.0
alembic==1.9.0
pytest==7.2.0
pytest-asyncio==0.20.3
httpx==0.23.3
pre-commit==2.21.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
bcrypt==4.0.1
greenlet==2.0.2
sentry-sdk[fastapi]
starlette-exporter==0.15.1
Faker==22.4.0
# starlette~=0.22.0
prometheus-fastapi-instrumentator==6.1.0
fastapi-cache2[redis]==0.2.1
celery==5.3.6
flower==2.0.1
redis==4.6.0
async-timeout==4.0.3 #E   ModuleNotFoundError: No module named 'async_timeout'
python-dotenv==1.0.0
